WORKING PROJECT -- need to fix GIFs and node_module
GIFs are not appearing in the story on the app. Probably not a browser
thing but double check. Most likely the page is unable to find the file
structure - is there a specific line we're forgetting to include?

Also, node_module recursive dependencies in the bin folder are acting up
again. Don't remember the solution we used last time to clear them; had
to make another server project. This is a known issue but requires more
research on how to effectively mitigate.